rename folder "toolContainer" -> "tools"
