+++ src/modules/toolbox/components/toolContainer/ToolContainer.js
@@ -4,7 +4,7 @@ import { DrawToolContent } from '../drawToolContent/DrawToolContent';
 import { MeasureToolContent } from '../measureToolContent/MeasureToolContent';
 import { ShareToolContent } from '../shareToolContent/ShareToolContent';
 import css from './toolContainer.css';
-import { setContainerContent, ToolKey } from '../../../../store/toolContainer/toolContainer.action';
+import { setContainerContent, ToolKey } from '../../../../store/tools/toolContainer.action';
 import { unsafeHTML } from 'lit-html/directives/unsafe-html.js';
 import closeIcon from './assets/x-square.svg';
 import { MvuElement } from '../../../MvuElement';
