+++ src/plugins/DrawPlugin.js
@@ -1,7 +1,7 @@
 import { observe } from '../utils/storeUtils';
 import { addLayer, removeLayer } from '../store/layers/layers.action';
 import { BaPlugin } from './BaPlugin';
-import { ToolKey } from '../store/toolContainer/toolContainer.action';
+import { ToolKey } from '../store/tools/toolContainer.action';
 import { activate, deactivate } from '../store/draw/draw.action';
 
 /**
