+++ src/services/StoreService.js
@@ -10,7 +10,7 @@ import { notificationReducer } from '../store/notifications/notifications.reduce
 import { createMediaReducer } from '../store/media/media.reducer';
 import { topicsContentPanelReducer } from '../store/topicsContentPanel/topicsContentPanel.reducer';
 import { modalReducer } from '../store/modal/modal.reducer';
-import { toolContainerReducer } from '../store/toolContainer/toolContainer.reducer';
+import { toolContainerReducer } from '../store/tools/toolContainer.reducer';
 import { drawReducer } from '../store/draw/draw.reducer';
 import { sharedReducer } from '../store/shared/shared.reducer';
 import { geolocationReducer } from '../store/geolocation/geolocation.reducer';
