+++ test/store/tools/toolContainer.reducer.test.js
@@ -1,6 +1,6 @@
 import { TestUtils } from '../../test-utils.js';
-import { toolContainerReducer } from '../../../src/store/toolContainer/toolContainer.reducer.js';
-import { setContainerContent } from '../../../src/store/toolContainer/toolContainer.action.js';
+import { toolContainerReducer } from '../../../src/store/tools/toolContainer.reducer.js';
+import { setContainerContent } from '../../../src/store/tools/toolContainer.action.js';
 
 
 describe('toolContainerReducer', () => {
